moving spark jobserver api classes to separate folder
