+++ src/main/java/com/qubole/qds/sdk/java/client/QdsClient.java
@@ -15,13 +15,13 @@
  */
 package com.qubole.qds.sdk.java.client;
 
-import com.qubole.qds.sdk.java.api.AppApi;
 import com.qubole.qds.sdk.java.api.ClusterApi;
 import com.qubole.qds.sdk.java.api.CommandApi;
 import com.qubole.qds.sdk.java.api.DbTapApi;
 import com.qubole.qds.sdk.java.api.HiveMetadataApi;
 import com.qubole.qds.sdk.java.api.ReportApi;
 import com.qubole.qds.sdk.java.api.SchedulerApi;
+import com.qubole.qds.sdk.java.api.sparkjobserver.AppApi;
 import com.qubole.qds.sdk.java.details.RequestDetails;
 import com.qubole.qds.sdk.java.details.ForPage;
 import javax.ws.rs.client.InvocationCallback;
