+++ src/main/java/com/qubole/qds/sdk/java/details/sparkjobserver/AppApiImpl.java
@@ -13,23 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.qubole.qds.sdk.java.details;
+package com.qubole.qds.sdk.java.details.sparkjobserver;
 
 import java.util.List;
 import javax.ws.rs.core.GenericType;
-import com.qubole.qds.sdk.java.api.AppApi;
-import com.qubole.qds.sdk.java.api.AppBuilder;
 import com.qubole.qds.sdk.java.api.InvokableBuilder;
+import com.qubole.qds.sdk.java.api.sparkjobserver.AppApi;
+import com.qubole.qds.sdk.java.api.sparkjobserver.AppBuilder;
 import com.qubole.qds.sdk.java.client.QdsClient;
-import com.qubole.qds.sdk.java.entities.App;
-import com.qubole.qds.sdk.java.entities.AppItem;
+import com.qubole.qds.sdk.java.details.GenericInvokableBuilderImpl;
+import com.qubole.qds.sdk.java.details.RequestDetails;
 import com.qubole.qds.sdk.java.entities.Message;
+import com.qubole.qds.sdk.java.entities.sparkjobserver.App;
+import com.qubole.qds.sdk.java.entities.sparkjobserver.AppItem;
 
 public class AppApiImpl implements AppApi
 {
     private final QdsClient client;
 
-    AppApiImpl(QdsClient client)
+    public AppApiImpl(QdsClient client)
     {
         this.client = client;
     }
